# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Personal Operations Manager",
    "description": "A private workspace with review-before-save agent proposals.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/chat": {
      "get": {
        "tags": [
          "Assistant"
        ],
        "summary": "History",
        "operationId": "history_api_v1_chat_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ChatResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid CSRF token or origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict or stale version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ]
      },
      "post": {
        "tags": [
          "Assistant"
        ],
        "summary": "Chat",
        "operationId": "chat_api_v1_chat_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid CSRF token or origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict or stale version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Use the csrf_token returned by /auth/session. Never store the session cookie in JavaScript."
          }
        ]
      },
      "delete": {
        "tags": [
          "Assistant"
        ],
        "summary": "Clear History",
        "operationId": "clear_history_api_v1_chat_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid CSRF token or origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict or stale version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Use the csrf_token returned by /auth/session. Never store the session cookie in JavaScript."
          }
        ]
      }
    },
    "/api/v1/chat/{run_id}/apply": {
      "post": {
        "tags": [
          "Assistant"
        ],
        "summary": "Apply Proposal",
        "operationId": "apply_proposal_api_v1_chat__run_id__apply_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Use the csrf_token returned by /auth/session. Never store the session cookie in JavaScript."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid CSRF token or origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict or stale version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ]
      }
    },
    "/api/v1/chat/{run_id}/discard": {
      "post": {
        "tags": [
          "Assistant"
        ],
        "summary": "Discard Proposal",
        "operationId": "discard_proposal_api_v1_chat__run_id__discard_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Use the csrf_token returned by /auth/session. Never store the session cookie in JavaScript."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid CSRF token or origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict or stale version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentTraceStep": {
        "properties": {
          "step": {
            "type": "string",
            "title": "Step"
          },
          "details": {
            "type": "string",
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "step",
          "details"
        ],
        "title": "AgentTraceStep"
      },
      "ChatRequest": {
        "properties": {
          "user_input": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "User Input"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_input",
          "request_id"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "user_input": {
            "type": "string",
            "title": "User Input"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "intent": {
            "type": "string",
            "enum": [
              "planner",
              "notes",
              "calendar",
              "reminder",
              "general"
            ],
            "title": "Intent"
          },
          "mode": {
            "type": "string",
            "enum": [
              "demo",
              "vertex"
            ],
            "title": "Mode"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "applied",
              "discarded",
              "info"
            ],
            "title": "Status"
          },
          "proposal": {
            "$ref": "#/components/schemas/Proposal"
          },
          "trace": {
            "items": {
              "$ref": "#/components/schemas/AgentTraceStep"
            },
            "type": "array",
            "title": "Trace"
          },
          "applied_ids": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Applied Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "created_at",
          "updated_at",
          "version",
          "user_input",
          "request_id",
          "intent",
          "mode",
          "model",
          "status",
          "proposal",
          "trace",
          "applied_ids"
        ],
        "title": "ChatResponse"
      },
...
      "EventCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "start_time",
          "end_time"
        ],
        "title": "EventCreate"
      },
...
      "Page_ChatResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChatResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ChatResponse]"
      },
...
      "Proposal": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 3000,
            "minLength": 1,
            "title": "Message"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskCreate"
            },
            "type": "array",
            "maxItems": 8,
            "title": "Tasks"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/SuggestedNote"
            },
            "type": "array",
            "maxItems": 1,
            "title": "Notes"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventCreate"
            },
            "type": "array",
            "maxItems": 4,
            "title": "Events"
          },
          "reminders": {
            "items": {
              "$ref": "#/components/schemas/ReminderCreate"
            },
            "type": "array",
            "maxItems": 4,
            "title": "Reminders"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "Proposal"
      },
...
      "ReminderCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "urgency": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Urgency",
            "default": "medium"
          },
          "suggestion": {
            "type": "string",
            "maxLength": 500,
            "title": "Suggestion",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "ReminderCreate"
      },
...
      "SuggestedNote": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "content": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Content"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 3000
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "action_items": {
            "items": {
              "type": "string",
              "maxLength": 200,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 8,
            "title": "Action Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "title": "SuggestedNote"
      },
...
      "TaskCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Priority",
            "default": "medium"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "TaskCreate"
      },
...
      "SafeError": {
        "type": "object",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "loc": {
                  "type": "array",
                  "items": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "integer"
                      }
                    ]
                  }
                },
                "msg": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "CookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "aiops_session"
      }
    }
  }
}